Add logger transmission from the Factory to the created child. Also add a new lightweight logger which catches standard information once convergence has been reached.
Some minor fixes were also done. The logger transmission from the Factory to the
created child means that all products must also be `Loggable`, and this has been
done when it was not the case.
A major fix was changing the logging calls from inside the Updater to all the
`check` method calls which allow to transmit the logging calls to the
subcriterion when using the Combined criterion, among other uses.