Fix CMake ALIAS IMPORTED target error
CMake allows aliasing from IMPORTED targets only in 3.11+, while
we want to support CMake 3.9+.
We used the aliasing to allow third party packages to be either
downloaded and build ourselves, or found through the `find_package`
command without any other changes in the code.

In order to do the same thing with 3.9+, the IMPORTED target is now
directly created with the expected name, which can also be a
namespaced target. Therefore, no ALIAS is necessary any more.