Fix the Jacobi compilation time.
This is an approach at fixing the Jacobi kernels compilation which aims at
sacrificing as little as possible the runtime performance and get down the total
compilation time for this file as much as possible, all the while having a way
to enable the full performance optimizations in an easy way.

This approach only touches the `generate` kernels in the end, and provides the
full performance for the `apply` and other kernels as they are fast enough to
compile.

+ Split the Jacobi kernels into multiple files for parallel compilation and
   smaller code size.
+ Tone down some optimizations, namely use `noinline` for one function and
   `#pragma unroll 1` in two places to prevent unrolling. This impacts the
   Jacobi `generate` kernels only.
+ Add a compilation flag for enabling back the full optimizations.

__NOTE:__ the generate kernel compilation with full optimizations still takes
above 30 minutes on my laptop for one architecture (Maxwell). Without the
optimizations enabled, it takes less than 3 minutes.