Fixed GKO_ENABLE_BUILD_METHOD macro
`GKO_ENABLE_BUILD_METHOD` takes an argument, which is supposed to be
the name of the factory class. However, this parameter was never used.
Instead, the name `Factory` was hard-coded, which worked because we
always use this name. However, it failed to compile when using a
different name.
Now, the argument is used properly, so names other than `Factory`
will also work.