Add a new example called `custom-logger`.
The purpose of this example is to show the users how to customize Ginkgo by
adding a new logger, which is useful and more efficient for application specific
problems. This is also one of the most basic (and simple) ways of customizing
Ginkgo, therefore this is a good entry level example.

This example simply prints a table of the recurrent residual norm against the
real residual norm.

This example is documented as much as well as the `simple-solver` example for
the user's convenience.