Fix version bugs with Intel ICC 2017.
The main bug is that this compiler is not able to construct this object without
explicitly declaring this constructor for both the version class and the storage
scheme class. In addition the version class constructor should be constexpr so
that `version` becomes a literal type which can be returned by a constexpr
function.

Also fix a unique_ptr to const vs unique_ptr to non const cast problem which
Intel 2017 cannot solve, and another unique_ptr of gko::LinOp vs actual
MatrixType.