Changed loop order in OpenMP COO SpMV
Since the previous version contained race conditions, the order of the
for-loops was swapped (first over number of right hand sides, then
matrix values). For a single right hand side, it performs close to the
reference implementation, while for multiple right hand sides, it is
faster in most cases (at least in the ones I tested).