Make ILU usable as a preconditioner
All preconditioner must have a factory of type `LinOpFactory` in order
to get accepted as a preconditioner in any solver. This is now possible
with ILU (a test was added to ensure that).
However, as part of that adaptation, it is no longer possible to use
`generate` with both L and U separately. Now, either a matrix can be
passed, which will be factorized with ParIlu internally, or a
composition can be passed, consisting of both L and U (in that order).