Add `generated_preconditioner` to all solvers
This allows us to generate a preconditioner once and use it in multiple
solvers without the need to generate it again.
Providing a factory is still possible as it was.
However, if both a factory and a generated preconditioner is given,
the generated preconditioner is used, and the factory ignored.