Improve isfinite function and the test for it
`isFinite()` was previously hand-implemented with a bit-mask since it
seemed to be the least amount of work. However, the PTX code that
`isfinite(x)` produces does `abs(x) < INF`.
Now, that is also what the hand-written implementation does.
Tests were also adjusted to always cover the hand-written
implmenentation as well as the `gko::isfinite()`.