Tune and benchmark variables non intrusively.
This piece of code integrates and shows how to tune some variables,
namely the multiple used to decide on the number of warps used in COO.

How: This is integrated into the SPMV benchmarks for now and is entirely
controlled with the new CMake option `GINKGO_BENCHMARK_ENABLE_TUNING`.

The setup works through the use of (dynamic) global variables so that
tuning is as little intrusive as possible. This should work even with the
"implementation selection" since this also has a dynamic component which
could use the same strategy as shown here.

Caveat: With this setup, only benchmarks can be compiled, other
executable codes such as tests would require to also add
`benchmark/utils/tuning_variables.cpp` or equivalent. This option should
of course not be turned on for general Ginkgo usage, but only for
tuning.