Make the Ginkgo-overhead example use `b=1`.
There is no impact of `b` value on the results. For example using the
ReferenceExecutor on these processors:
Intel(R) Xeon(R) Gold 6248R CPU
With Nan:
Running 1000000 iterations of the CG solver took a total of 1.28285 seconds.
        Average library overhead:     1282.85 [nanoseconds / iteration]

With 1.0:
Running 1000000 iterations of the CG solver took a total of 1.26713 seconds.
        Average library overhead:     1267.13 [nanoseconds / iteration]