Move row_major and block_col_major to accessor/
Only the new multidimensional row_major is moved, the 2D row_major that
is part of the public interface remains.
Additionally, unnecessary code from public accessors was removed.
The code that was removed was never in the release and originally
intended for usage inside the accessors. Since they moved to an
independent directory, that code is no longer necessary.
This removal prevents it from ever becomming part of the public
interface.