HIP overhaul, Hwloc and test_install fixes.
On hwloc
________

There were a few issues with the CMake hwloc setup.
+ hwloc was always built even when a valid version already existed on
the system.
+ The FindHWLOC module had a few issues in terms of organization, search
paths as well as variable management. This has been fixed.
+ hwloc is a public dependency of Ginkgo but was not installed when we
use the one from our build directory, that would create errors with
missing libraries in some cases.
+ The `HWLOC_XMLFILE` should not be managed by Ginkgo, hwloc already
takes care of it, if the environment variable `HWLOC_XMLFILE` is set, it
uses that as base, otherwise it uses the machine topology it can
detect.
+ The core/devices which are always compiled were moved into their own
root directory. This better reflects the new library that it creates as
well as the fact all libraries depend on it.

On test install
________

Test install was missing a HIP version which I added. In addition, the
commands were not properly triggered thus only the base (non CUDA
version) was tested, that is now fixed.

On RPATH management
_________

For HIP and HWLOC, the INSTALL_RPATH property is now properly managed.
Since all libraries require HWLOC, this setting is added directly into
`ginkgo_install_libraries`. For HIP, the setting is added only when
actually building HIP.

On HIP
________

This part saw a lot of changes.

+ ROCm < 3.5 support is dropped as it is now old and all HPC clusters
always update the ROCm version since they know the ecosystem is somewhat
unstable.
+ Instead of linking against `hip::device` in `ginkgo_hip`, I link
against `libamdhip64.so` publicly.
+ In general, the purpose of this rework is to depend more on CMake
transitive properties such that we remove special cases. Thanks to this
and the previous point as well, it is possible to remove entirely the
`hip_helpers.cmake` file and remove all of the bans we were doing.
+ Thanks to relying on transitive properties, there is no need for the
`special_linkage` test case anymore.
+ Instead of building the `RPATH_FOR_HIP` chain, in Ginkgo everything
uses the `CXX` language for linking, which properly takes into account
RPATH information unlike when liking with the `HIP` language. For users,
a simple string `GINKGO_INSTALL_RPATH_FOR_HIP` is provided which is a
linking flag to be given to the `HIP` language linker after installing
Ginkgo.
+ In general, there is also better management or relying more on the
defaults for some environment variables and HIP variables.

Open issues
________

An annoying issue (still), although it's mitigated thanks to linking
publicly against `libamdhip64.so` which doesn't create inconvenience to
the user, is that we have to include `hip/hip_runtime.hpp` in our public
headers (to find `__forceinline__` as well as `HIP_DEVICE_COMPILE`, see
https://github.com/ginkgo-project/ginkgo/blob/develop/include/ginkgo/core/base/types.hpp#L46-L85.
In addition, the `GKO_ASSERT` cannot be reworked as well otherwise it
creates mysterious issues in the `check_headers` test, see:
https://gitlab.com/ginkgo-project/ginkgo-public-ci/-/jobs/1073846520