Unified batch entry creation functions for omp and reference
There were two identical copies of these functions in different
directories. Now, the reference version is in a 'host' namespace instead
of 'reference', and is called by both reference and omp.

Eventually, we could move to this kind of scheme for 'device'-level
implementations of solver kernels, instead of the current hpp.inc files.