checks benchmark completion in each iteration
This reworks the previous adaptiv benchmarking.
Now, the number of iteration is determined on-the-fly instead of before-hand.
Also new cmd-line flags have been added to allow for greater control over the adaptive benchmarks.
The usage is similar to google benchmark.

Currently, it is not possible to use this approach for the preconditioner benchmark, as these do not update the runtime in each iteration.