change clang-format PointerAlignment to Left
This avoids the clang-format inconsistencies related to variadic
pointer and reference parameters with PointerAlignment: Right.
Older clang-format versions format `T&&... a` as `T &&... a`,
but never versions format it as `T &&... a`, while normal
variadic parameters are formatted as `T... a` any time.
This inconsistency is not present for PointerAlignment: Left