use separate zero init guess
this is necessary to distinguish between a user supplied initial guess and
the default zero initial guess. if the user supplies an initial guess, it's
assumed that the guess contains the current constrained values. if the user
doesn't specify an initial guess, every time the constrained values are
updated, the default initial guess has to be updated as well.