remove partition member from vector
for now, no functionality (except reading/consistency checking) depends on the partition, so there is no point in storing it. Regarding the consistency check: we can ignore that for the moment, since we have to use a pointer comparison to check for equality, which is not sufficient for the general use case.