changes default coarse grid solver
Before:
  identity
Now:
  GMRES with Jacobi, solved to machine precision*10 and max. number of iterations = size(matrix)
Rational:
The coarse grid solve usually should be exact (up to machine precision). With GMRES we should be able to achieve that for the largest set of matrices. Since we would expect the matrices to be very small, a scalar Jacobi preconditioner should be sufficient, compared to alternatives like ILU.