don't use MPI from compiler wrapper directly
Using MPI from the MPI compiler wrapper directly means that no include path or library is added for MPI. This gives errors if a file that is not compiled by the wrappers and is linked against/includes ginkgo can't find the MPI headers or implementation. This mostly effects device code, since they will most likely not be compiled with the MPI wrapper compilers.