only mirror when more than documentation change
Note. Github action need to fulfill path and branch condition at the same time
Github does not support two push in the same file.
We split it into two file such that we always mirror on the major branches but only mirror when more than doc change in the other branches