Show correct line numbers in generated source lines for common tests
For CUDA/HIP and DPCPP, we create separate copies of the source files using configure_file(...)
to apply different CMake file properties to them.
For the compiler to report error messages with the correct source file name,
this adds a #line preprocessor statement to the top of these generated files.